+++ src/command/ECommand.ts
@@ -1,4 +1,5 @@
 export const enum ECommand {
+    CmdFindClassRef = 'ahk.nekoHelp.CmdFindClassRef',
     CmdFindFuncRef = 'ahk.nekoHelp.CmdFindFuncRef',
     CmdGotoFuncDef = 'ahk.nekoHelp.CmdGotoFuncDef',
     CompletionMsgBox = 'ahk.nekoHelp.CompletionMsgBox', // TODO